requirements.txt
# Core Django
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-mptt==0.14.0

# Database
psycopg2-binary==2.9.7
django-extensions==3.2.3

# Authentication & Authorization
django-oauth-toolkit==1.7.1
social-auth-app-django==5.4.0
PyJWT==2.8.0

# API Documentation
drf-spectacular==0.26.5

# Environment variables
python-decouple==3.8

# SMS Integration (Africa's Talking)
africastalking==1.2.5

# Email
django-anymail==10.2

# Testing
pytest==7.4.3
pytest-django==4.5.2
pytest-cov==4.1.0
factory-boy==3.3.0

# Development
django-debug-toolbar==4.2.0
black==23.10.1
flake8==6.1.0
isort==5.12.0

# Deployment
gunicorn==21.2.0
whitenoise==6.6.0

# Container health checks
django-health-check==3.17.0